refactor(web): extract backtesting form actions
